requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.2
django-allauth==0.49.0
django-crispy-forms==1.14.0
filelock==3.4.2
fonttools==4.29.1
gunicorn==20.1.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pipenv==2021.11.23
platformdirs==2.4.1
protobuf==3.19.4
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tensorflow-hub==0.12.0
tqdm==4.63.0
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
whitenoise==6.0.0
